removing <Spectacle/> tag
