+++ package.json
@@ -27,6 +27,7 @@
     "react-typography": "^0.15.0",
     "redux": "^3.0.4",
     "redux-actions": "^0.12.0",
+    "styled-components": "^1.1.3",
     "tween-functions": "1.1.0",
     "victory-core": "^9.1.1"
   },
