+++ src/index.js
@@ -16,7 +16,6 @@ import Markdown from "./components/markdown";
 import Quote from "./components/quote";
 import S from "./components/s";
 import Slide from "./components/slide";
-import Spectacle from "./components/spectacle";
 import TableHeaderItem from "./components/table-header-item";
 import TableItem from "./components/table-item";
 import TableRow from "./components/table-row";
@@ -43,7 +42,6 @@ export {
   Quote,
   S,
   Slide,
-  Spectacle,
   TableHeaderItem,
   TableItem,
   TableRow,
